Move stuff around to better separate compile-time and runtime types and functions
