+++ site/frontend/src/pages/compare/runtime/common.ts
@@ -1,0 +1,14 @@
+import {BenchmarkFilter, StatComparison} from "../types";
+
+export type RuntimeBenchmarkFilter = BenchmarkFilter;
+
+export const defaultRuntimeFilter: RuntimeBenchmarkFilter = {
+  name: null,
+  nonRelevant: false,
+  showRawData: false,
+};
+
+export interface RuntimeBenchmarkComparison {
+  benchmark: string;
+  comparison: StatComparison;
+}
